refactor: split work session command family
